cmake: make use of pkg-tool .pc for dvdnav, avcodec, dvdread time.h: force "struct timespec" in time.h as internal issues stop defining this structure with this gcc to date. libdvdread-NOW: try to  bypass the compiling of this directory.

+++ src/backup/k9dvdbackup.cpp
@@ -907,7 +907,7 @@ k9Vobu * k9DVDBackup::remapOffset(uint32_t _sector,uint32_t *_offset,int _dir) {
 
 
         if ((vobu1 !=NULL) && (vobu2!=NULL)) {
-            *_offset = abs(vobu1->newSector - vobu2->newSector)  | maskOffset1 ;
+            *_offset = abs((long)vobu1->newSector - vobu2->newSector)  | maskOffset1 ;
             *_offset |= maskOffset2;
             return vobu2;
         }
